Expressions like "ein Hotel" or "ein Angebot" in German are not 1h (#190)
* Expressions like "ein Hotel" or "ein Angebot" in German are not 1h; fix
this by adding some minimal requirement on the context, i.e. that there
is a space after the "H" or "A" or that this appears at the end of the
text. Actuall the rule for "am|pm" is pretty useless for German as "am|pm" is at
most a corner case.

* generalized token expression allowed as a follow up to all punctation
chars, all letters and end of text