[EN] More natural disambiguation of intervals
This commit adds two new rules to "better" handle ambiguous time intervals
suffixed with an AM or PM. Typically, "from 7 to 8 PM" and "between 7 and 8 PM"
refer to the interval from 1900 to 2000, and _not_ the interval 0700 to 2000.

This behavior has been a bit annoying to end users of our application. I recognize this
change represents an opinionated design decision, but I do think my proposed semantics
more mainstream.

This is my first time writing Clojure since undergraduate, so please do let me know
if anything is particularly cringeworthy ... This is also my excuse for not refactoring
the AM PM bit into a helper function, though I would happily take a stab at that change
if required!